[ANDROSDK-2125] refactor custom intent stores and handlers to match table name
