+++ core/src/main/java/org/hisp/dhis/android/core/settings/internal/CustomIntentDataElementStore.kt
@@ -28,32 +28,8 @@
 
 package org.hisp.dhis.android.core.settings.internal
 
-import android.database.Cursor
-import org.hisp.dhis.android.core.arch.db.access.DatabaseAdapter
-import org.hisp.dhis.android.core.arch.db.stores.binders.internal.StatementBinder
-import org.hisp.dhis.android.core.arch.db.stores.binders.internal.StatementWrapper
-import org.hisp.dhis.android.core.arch.db.stores.internal.LinkStoreImpl
-import org.hisp.dhis.android.core.settings.CustomIntentAttribute
-import org.hisp.dhis.android.core.settings.CustomIntentAttributeTableInfo
-import org.koin.core.annotation.Singleton
+import org.hisp.dhis.android.core.arch.db.stores.internal.LinkStore
+import org.hisp.dhis.android.core.settings.CustomIntentDataElement
 
-@Singleton
-@Suppress("MagicNumber")
-internal class CustomIntentAttributeTriggerStoreImpl(
-    databaseAdapter: DatabaseAdapter,
-) : CustomIntentAttributeTriggerStore,
-    LinkStoreImpl<CustomIntentAttribute>(
-        databaseAdapter,
-        CustomIntentAttributeTableInfo.TABLE_INFO,
-        CustomIntentAttributeTableInfo.Columns.CUSTOM_INTENT_UID,
-        BINDER,
-        { cursor: Cursor -> CustomIntentAttribute.create(cursor) },
-    ) {
-
-    companion object {
-        private val BINDER = StatementBinder { o: CustomIntentAttribute, w: StatementWrapper ->
-            w.bind(1, o.uid())
-            w.bind(2, o.customIntentUid())
-        }
-    }
-}
+@Suppress("EmptyClassBlock")
+internal interface CustomIntentDataElementStore : LinkStore<CustomIntentDataElement>
